修复QQ号映射的问题(云黑写的qq号type是int) 修复提示信息乱码的问题

+++ src/main/java/xyz/blackbe/blackbeplugin/data/BlackBEBlacklistCheckData.java
@@ -84,7 +84,7 @@ public class BlackBEBlacklistCheckData extends BlackBEData {
             private String xuid;
             private String info;
             private Integer level;
-            private Integer qq;
+            private Long qq;
 
             public String getUUID() {
                 return uuid;
@@ -110,7 +110,7 @@ public class BlackBEBlacklistCheckData extends BlackBEData {
                 return level;
             }
 
-            public Integer getQQ() {
+            public Long getQQ() {
                 return qq;
             }
